requirements.txt
whatsapp-python
fastapi
uvicorn
requests_toolbelt